vm: toggle network based on systemd-networkd presence in cpio
Add a _rt_cpio_has helper to determine whether a path is carried in the
first cpio archive of an initramfs image. Use the presence of
systemd-networkd (added via _rt_require_networking) when determining
whether or not to assign a tap device to the VM, instead of the previous
"user.rapido.vm_networkless" xattr.

Signed-off-by: David Disseldorp <ddiss@suse.de>