readme: rework and reformat as markdown
In preparation for adding an SVG render of the arch diagram, it seems
that markdown would be a better option, as I can embed the SVG image
alongside the ASCII art without cluttering up the page too much on both
web and text version.

This change also collapses some of the dependency duplication in the
overview / setup sections.

The architecture ASCII art is modified to better suit svgbob, which
will be used for SVG rendering.

Signed-off-by: David Disseldorp <ddiss@suse.de>